Moved changeMobile function to navbar component
